Sprint 14 project completed

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -10,9 +10,9 @@ public interface UserService {
 
     List<User> getAllUsers();
 
-    User getUserById(Integer userId);
+    User getUserById(Long userId);
 
-    User updateUser(Integer userId, User user);
+    User updateUser(Long userId, User user);
 
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
 }
